libcusmm/kernels: rename scripts to more descriptive names
