+++ src/acc/libsmm_acc/libcusmm/kernels/cusmm_dnt_tiny.py
@@ -10,7 +10,7 @@
 
 
 import numpy as np
-from kernels.cusmm_dnt import Kernel, round_up_to_nearest_multiple
+from kernels.cusmm_dnt_base import Kernel, round_up_to_nearest_multiple
 
 
 class Kernel_dnt_tiny(Kernel):
