+++ src/acc/libsmm_acc/libcusmm/predict_helpers.py
@@ -15,7 +15,7 @@ import pickle
 import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
-from kernels.cusmm_dnt_helper import to_string
+from kernels.cusmm_predict import to_string
 
 
 # ===============================================================================
