+++ src/acc/libsmm_acc/libcusmm/tune_merge.py
@@ -12,7 +12,7 @@
 
 import json
 from optparse import OptionParser
-from kernels.cusmm_dnt_helper import params_dict_to_kernel
+from kernels.cusmm_predict import params_dict_to_kernel
 
 
 def main():
